Refactoring
- bring back SourcePath
- SourceLocator has many different handlers for different sources
