+++ lib/reek/cli/input.rb
@@ -7,7 +7,7 @@ module Reek
     #
     module Input
       def sources
-        Source::SourceLocator.build(from_argv || from_pipe || from_working_directory)
+        Source::SourceLocator.new(from_argv || from_pipe || from_working_directory).call
       end
 
       private
