+++ lib/reek/source/source_code.rb
@@ -4,7 +4,7 @@ Reek::CLI::Silencer.silently do
 end
 require_relative '../tree_dresser'
 require_relative '../ast/node'
-require_relative '../source/source_locator'
+require_relative '../source/source_path'
 
 module Reek
   module Source
@@ -44,7 +44,7 @@ module Reek
         when IO            then new(source.readlines.join, IO_IDENTIFIER)
         when Pathname      then new(source.read, source.to_s)
         when String        then new(source, STRING_IDENTIFIER)
-        when SourceLocator then new(source.read, source.to_s)
+        when SourcePath then new(source.read, source.to_s)
         end
       end
