+++ spec/reek/examiner_spec.rb
@@ -43,7 +43,7 @@ RSpec.describe Reek::Examiner do
     let(:examiner) { described_class.new(smelly_file, [], configuration: configuration) }
     let(:path) { SAMPLES_PATH.join('all_but_one_masked/masked.reek') }
     let(:smelly_file) do
-      Reek::Source::SourceLocator.new(Pathname.glob(SAMPLES_PATH.join('all_but_one_masked/d*.rb')).first)
+      Reek::Source::SourcePath.new(Pathname.glob(SAMPLES_PATH.join('all_but_one_masked/d*.rb')).first)
     end
 
     it_should_behave_like 'one smell found'
@@ -51,7 +51,7 @@ RSpec.describe Reek::Examiner do
 
   context 'with a fragrant File' do
     let(:clean_file) do
-      Reek::Source::SourceLocator.new(Pathname.glob(SAMPLES_PATH.join('three_clean_files/*.rb')).first)
+      Reek::Source::SourcePath.new(Pathname.glob(SAMPLES_PATH.join('three_clean_files/*.rb')).first)
     end
     let(:examiner) { described_class.new(clean_file) }
